backend/requirements.txt
flask
pymongo
bcrypt
flask-socketio
torch 
scikit-learn
flask_cors
pandas
python-dotenv
requests